Move `Commands` function to command.go (no code change)
