+++ cmd/agent/subcommands/run/command.go
@@ -1,0 +1,59 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// Package run implements 'agent run' (and deprecated 'agent start').
+package run
+
+import (
+	"github.com/DataDog/datadog-agent/cmd/agent/command"
+	"github.com/DataDog/datadog-agent/cmd/agent/common/path"
+	"github.com/DataDog/datadog-agent/comp/core"
+	"github.com/DataDog/datadog-agent/comp/core/config"
+	"github.com/DataDog/datadog-agent/comp/core/log/logimpl"
+	"github.com/DataDog/datadog-agent/comp/core/secrets"
+	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig/sysprobeconfigimpl"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"github.com/spf13/cobra"
+	"go.uber.org/fx"
+)
+
+// Commands returns a slice of subcommands for the 'agent' command.
+func Commands(globalParams *command.GlobalParams) []*cobra.Command {
+	cliParams := &cliParams{
+		GlobalParams: globalParams,
+	}
+	runE := func(*cobra.Command, []string) error {
+		// TODO: once the agent is represented as a component, and not a function (run),
+		// this will use `fxutil.Run` instead of `fxutil.OneShot`.
+		return fxutil.OneShot(run,
+			fx.Supply(cliParams),
+			fx.Supply(core.BundleParams{
+				ConfigParams:         config.NewAgentParams(globalParams.ConfFilePath),
+				SecretParams:         secrets.NewEnabledParams(),
+				SysprobeConfigParams: sysprobeconfigimpl.NewParams(sysprobeconfigimpl.WithSysProbeConfFilePath(globalParams.SysProbeConfFilePath)),
+				LogParams:            logimpl.ForDaemon(command.LoggerName, "log_file", path.DefaultLogFile),
+			}),
+			getSharedFxOption(),
+			getPlatformModules(),
+		)
+	}
+
+	runCmd := &cobra.Command{
+		Use:   "run",
+		Short: "Run the Agent",
+		Long:  `Runs the agent in the foreground`,
+		RunE:  runE,
+	}
+	runCmd.Flags().StringVarP(&cliParams.pidfilePath, "pidfile", "p", "", "path to the pidfile")
+
+	startCmd := &cobra.Command{
+		Use:        "start",
+		Deprecated: "Use \"run\" instead to start the Agent",
+		RunE:       runE,
+	}
+	startCmd.Flags().StringVarP(&cliParams.pidfilePath, "pidfile", "p", "", "path to the pidfile")
+
+	return []*cobra.Command{startCmd, runCmd}
+}
